Extract CodeBuild trigger Lambda to a linted module
Move the inline build-trigger custom-resource handler out of agent_stack.py
into agent/cdk/build_trigger/index.py, loaded via Code.from_asset, so it is
covered by ruff, formatting, and static analysis (the handler was previously
an opaque r-string invisible to linting). Behavior is unchanged: start the
CodeBuild build, poll to completion, and gate the AgentCore Runtime on the
image being in ECR. Addresses reviewer feedback.
